enable existing TextLineTest for all targets
